requirements.txt
pandas
GitPython
termcolor